refactor: GrassCard constant 분리
