+++ src/constants/GrassCard/GrassCard.ts
@@ -1,0 +1,12 @@
+export const GRASS_TITLE = {
+  USER_PROGRESS_MESSAGE: "님은 지금까지\n",
+  TOTAL_DAYS: "총 ",
+  PLANTED_DAYS: "일의 잔디를 심으셨어요!"
+};
+
+
+export const GRASS_CARD_ALT_MESSAGE: {JANDI: 'jandi'} = {
+  JANDI: 'jandi',
+};
+
+
