--- src/constants/GrassCard/GrassCardAltMessage.ts
@@ -1,3 +1,0 @@
-export const GrassCardAltMessage: {JANDI: 'jandi'} = {
-  JANDI: 'jandi',
-};
